+++ src/hooks/useElection.ts
@@ -36,22 +36,12 @@ const useElection = () => {
   const electionOutcomes = React.useMemo(() => {
     if (election) {
       return {
-        fptp: election.fptp(),
-        veto: election.veto(),
-        signed: convertDetailedToSimple(election.signed()),
-        vfa: convertDetailedToSimple(election.vfa()),
         irv: election.irv(),
         coombs: election.coombs().map(convertDetailedToSimple),
         fab_irv: election.fab_irv().map(convertDetailedToSimple),
         contingency: election.contingency(),
         supplementary: election.supplementary(),
         sl_contingency: election.sl_contingency(),
-        borda: election.borda(),
-        nauru: election.nauru(),
-        eurovision: election.eurovision(),
-        dabagh: election.dabagh(),
-        approval: election.approval(),
-        cav: election.cav(),
       }
     } else return {};
   }, [election]);